refactor(viewmodel): extract shared post presenter from Listing and Detail
Listing and Detail each carried five identical presenter methods plus an
identical authorCache/loadAuthor pair. Extract them to PostPresenterTrait
behind PostPresenterInterface, which is the contract post/card.phtml consumes.

No behaviour change: the existing unit suite passes unmodified.
